# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de base en laravel",
        "description": "MDN",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://mdn-rdv.com/api"
        }
    ],
    "paths": {
...
        "/admin/societies/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__SOCIETIES"
                ],
                "summary": "Enregistre une société",
                "description": "",
                "operationId": "169a92719d4a68ecbf15285bf7601586",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Orange"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/societies/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__SOCIETIES"
                ],
                "summary": "Renvoie les informations d'une société par son ID",
                "operationId": "fb27adff1ad221ebaedbcd1d9253140a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/societies/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__SOCIETIES"
                ],
                "summary": "Editer les informations d'une société",
                "operationId": "426d1726589b176b19dd25248f4a03f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "libelle": {
                                        "type": "string",
                                        "example": "Radiologie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/societies/delete": {
            "post": {
                "tags": [
                    "ADMINISTRATION__SOCIETIES"
                ],
                "summary": "Supprimer une société",
                "operationId": "5b05a67d4f364489e32a26ff3bcfe607",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            },
            "0": []
        }
    }
}